+++ nodejs/awsx/apigateway/index.ts
@@ -13,9 +13,12 @@
 // limitations under the License.
 
 export * from "./api";
-export * from "./lambdaauthorizer";
+
+export * from "./apikey";
 export * from "./cognitoauthorizer";
+export * from "./lambdaAuthorizer";
 export * from "./metrics";
+export * from "./swagger_json";
 
 // @pulumi/awsx is a deployment-only module.  If someone tries to capture it, and we fail for some
 // reason we want to give a good message about what the problem likely is.  Note that capturing a